Initial typescript server-side conversion and restructure
linting
paths not working
blah
