+++ app/client/images/favicon.ico
