+++ app/server/app.js
@@ -1,4 +1,4 @@
-const appInsights = require('app-insights');
+const appInsights = require('app/server/app-insights');
 const { Express } = require('@hmcts/nodejs-logging');
 const nunjucks = require('nunjucks');
 const express = require('express');
@@ -6,7 +6,7 @@ const routes = require('app/routes');
 const errors = require('app/middleware/error-handler');
 const health = require('app/middleware/health');
 const locale = require('app/locale/en.json');
-const paths = require('paths');
+const paths = require('app/server/paths');
 const bodyParser = require('body-parser');
 
 const isDevelopment = process.env.NODE_ENV === 'development';
@@ -38,7 +38,7 @@ function setup(sessionHandler, options) {
     extended: true
   }));
 
-  app.use('/public', express.static(`${__dirname}/public`));
+  app.use('/public', express.static(`${__dirname}/../../public`));
 
   app.use(Express.accessLogger());
