+++ app/server/tsconfig.json
@@ -1,0 +1,21 @@
+
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "esModuleInterop": true,
+        "target": "es6",
+        "noImplicitAny": true,
+        "moduleResolution": "node",
+        "sourceMap": false,
+        "baseUrl": ".",
+        "paths": {
+            "*": [
+                "./*",
+                "node_modules/*"
+            ]
+        }
+    },
+    "include": [
+        "./server.ts"
+    ]
+}
