+++ package.json
@@ -7,14 +7,14 @@
     "node": ">=8.9.0 <9.0"
   },
   "scripts": {
-    "start": "cross-env NODE_PATH=. node server.js",
-    "dev": "yarn build && NODE_PATH=. NODE_ENV=development nodemon -e js,json,html server.js",
+    "start": "cross-env NODE_PATH=. node app/server/server.js",
+    "dev": "yarn build && NODE_PATH=. NODE_ENV=development nodemon -e js,json,html app/server/server.js",
     "dev-mock-api": "concurrently \"yarn dev\" \"yarn mock-api\"",
     "mock-api": "dyson test/mock/services/ 8080",
-    "build": "yarn build-sass && yarn build-js && yarn copy-assets",
-    "build-sass": "node-sass app/assets/sass/application.scss public/css/main.css --output-style compressed",
+    "build": "yarn build-sass && yarn build-js && yarn copy-assets && yarn tsc -p app/server/tsconfig.json",
+    "build-sass": "node-sass app/client/sass/application.scss public/css/main.css --output-style compressed",
     "copy-assets": "cp -R ./node_modules/govuk-frontend/assets ./public/govuk-frontend",
-    "build-js": "browserify ./app/assets/javascript/main.ts -p [ tsify --noImplicitAny ] -o ./public/js/main.js",
+    "build-js": "browserify ./app/client/javascript/main.ts -p [ tsify -p app/client --noImplicitAny ] -o ./public/js/main.js",
     "test": "yarn test:unit && yarn test:unit-ts && yarn test:browser",
     "test:unit": "cross-env NODE_PATH=. NODE_ENV=test mocha test/unit --recursive",
     "test:unit-ts": "cross-env NODE_PATH=. NODE_ENV=test mocha  -r ts-node/register test/unit/**/*.test.ts --recursive",
@@ -37,6 +37,7 @@
   "dependencies": {
     "@hmcts/nodejs-healthcheck": "^1.4.6",
     "@hmcts/nodejs-logging": "^3.0.0",
+    "@types/express": "^4.16.0",
     "applicationinsights": "^1.0.1",
     "body-parser": "^1.17.1",
     "config": "^1.30.0",
