+++ test/accessibility/runPa11y.js
@@ -4,7 +4,7 @@ const createSession = require('app/middleware/session');
 const { setup } = require('app');
 const pa11y = require('pa11y');
 const supertest = require('supertest');
-const paths = require('paths');
+const paths = require('app/server/paths');
 const dysonSetup = require('test/mock/dysonSetup');
 
 const app = setup(createSession(), { disableAppInsights: true });
