+++ test/browser/common.js
@@ -3,7 +3,7 @@ const puppeteer = require('puppeteer');
 const { createServer } = require('http');
 const createSession = require('app/middleware/session');
 const coh = require('test/fixtures/coh');
-const { setup } = require('app');
+const { setup } = require('app/server/app');
 const config = require('config');
 const dysonSetup = require('test/mock/dysonSetup');
