+++ test/browser/task-list.test.js
@@ -3,7 +3,7 @@ const { expect } = require('test/chai-sinon');
 const { startServices } = require('test/browser/common');
 const mockData = require('test/mock/services/allQuestions').template;
 const TaskListPage = require('test/page-objects/task-list');
-const paths = require('paths');
+const paths = require('app/server/paths');
 const config = require('config');
 
 const testUrl = config.get('testUrl');
