+++ test/unit/controllers/taskList.test.js
@@ -1,7 +1,7 @@
 const { expect, sinon } = require('test/chai-sinon');
 const { setupTaskListController, getTaskList } = require('app/controllers/taskList');
 const { INTERNAL_SERVER_ERROR } = require('http-status-codes');
-const appInsights = require('app-insights');
+const appInsights = require('app/server/app-insights');
 const express = require('express');
 
 describe('controllers/taskList.js', () => {
