+++ test/unit/paths.test.js
@@ -1,5 +1,5 @@
 const { expect } = require('test/chai-sinon');
-const paths = require('paths');
+const paths = require('app/server/paths');
 
 describe('paths.js', () => {
   it('should return an object', () => {
